ml/requirements.txt
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
onnx>=1.14.0
onnxruntime>=1.15.0
pyyaml>=6.0

# Optional for TFLite export
# onnx-tf>=1.10.0
# tensorflow>=2.13.0
